Move LinkedProof tests
