+++ sunscreen/Cargo.toml
@@ -52,6 +52,7 @@ criterion = { workspace = true }
 env_logger = { workspace = true }
 float-cmp = { workspace = true }
 lazy_static = { workspace = true }
+logproof = { workspace = true }
 proptest = { workspace = true }
 sunscreen_zkp_backend = { workspace = true, features = ["bulletproofs"] }
 sunscreen_compiler_common = { workspace = true }
